+++ notification/core/src/main/kotlin/vp/togedo/listener/GroupEventListener.kt
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import org.springframework.kafka.annotation.KafkaListener
 import org.springframework.stereotype.Component
 import vp.togedo.kafka.data.group.InviteGroupEventDto
-import vp.togedo.data.notification.EventEnums
-import vp.togedo.data.notification.SSEDao
+import vp.togedo.data.sse.EventEnums
+import vp.togedo.data.sse.SSEDao
 import vp.togedo.service.FCMService
 import vp.togedo.service.NotificationService
