+++ notification/presentation/src/main/kotlin/vp/togedo/controller/NotificationController.kt
@@ -5,7 +5,7 @@ import org.springframework.http.MediaType
 import org.springframework.http.codec.ServerSentEvent
 import org.springframework.web.bind.annotation.*
 import reactor.core.publisher.Flux
-import vp.togedo.data.notification.SSEDto
+import vp.togedo.data.sse.SSEDto
 import vp.togedo.service.NotificationService
